+++ lib/ckb/types/transaction.rb
@@ -12,7 +12,7 @@ module CKB
       # @param inputs [CKB::Types::Input[]]
       # @param outputs [CKB::Types::Output[]]
       # @param outputs_data [String[]]
-      # @param witnesses [CKB::Types::Witness[]]
+      # @param witnesses [String[]]
       def initialize(
         hash: nil,
         version: 0,
@@ -39,15 +39,14 @@ module CKB
         raise "Invalid number of witnesses!" if witnesses.length < inputs.length
 
         signed_witnesses = witnesses.map do |witness|
-          old_data = witness.data || []
+          old_datum = witness
           blake2b = CKB::Blake2b.new
           blake2b.update(Utils.hex_to_bin(tx_hash))
-          old_data.each do |datum|
-            blake2b.update(Utils.hex_to_bin(datum))
-          end
+          blake2b.update(Utils.hex_to_bin(old_datum))
           message = blake2b.hexdigest
-          data = [key.sign_recoverable(message)] + old_data
-          Types::Witness.from_h(data: data)
+          data = key.sign_recoverable(message) + old_datum[2..-1]
+
+          data
         end
 
         self.class.new(
@@ -62,6 +61,26 @@ module CKB
         )
       end
 
+      # @param index [Integer]
+      # @param key [CKB::Key]
+      #
+      # @return [CKBP::Types::Transaction]
+      def sign_input(index, key)
+        @hash = @hash || compute_hash
+
+        witness = witnesses[index] || ""
+
+        blake2b = CKB::Blake2b.new
+        blake2b.update(Utils.hex_to_bin(@hash))
+        blake2b.update(Utils.hex_to_bin(witness))
+        message = blake2b.hexdigest
+        signed_witness = key.sign_recoverable(message) + witness[2..-1]
+
+        witnesses[index] = signed_witness
+
+        self
+      end
+
       def to_h
         hash = {
           version: Utils.to_hex(@version),
@@ -70,7 +89,7 @@ module CKB
           inputs: @inputs.map(&:to_h),
           outputs: @outputs.map(&:to_h),
           outputs_data: @outputs_data,
-          witnesses: @witnesses.map(&:to_h)
+          witnesses: @witnesses
         }
         hash[:hash] = @hash if @hash
         hash
@@ -84,7 +103,7 @@ module CKB
           inputs: @inputs.map(&:to_h),
           outputs: @outputs.map(&:to_h),
           outputs_data: @outputs_data,
-          witnesses: @witnesses.map(&:to_h)
+          witnesses: @witnesses
         }
       end
 
@@ -106,7 +125,7 @@ module CKB
           inputs: hash[:inputs].map { |input| Input.from_h(input) },
           outputs: hash[:outputs].map { |output| Output.from_h(output) },
           outputs_data: hash[:outputs_data],
-          witnesses: hash[:witnesses].map { |witness| Witness.from_h(witness) }
+          witnesses: hash[:witnesses]
         )
       end
     end