+++ spec/ckb/rpc_spec.rb
@@ -1,8 +1,32 @@
+# frozen_string_literal: true
+
 RSpec.describe CKB::RPC do
   before do
     skip "not test rpc" if ENV["SKIP_RPC_TESTS"]
   end
-
+  let(:raw_block_h) do
+    { uncles: [],
+      proposals: [],
+      transactions: [{ version: "0x0",
+                       cell_deps: [],
+                       header_deps: [],
+                       inputs: [{ previous_output: { tx_hash: "0x0000000000000000000000000000000000000000000000000000000000000000", index: "0xffffffff" },
+                                  since: "0x1" }],
+                       outputs: [],
+                       outputs_data: [],
+                       witnesses: ["0x590000000c00000055000000490000001000000030000000310000009bd7e06f3ecf4be0f2fcd2188b23f1b9fcc88e5d4b65a8637b17723bbda3cce8011400000036c329ed630d6ce750712a477543672adab57f4c00000000"] }],
+      header: { compact_target: "0x20010000",
+                number: "0x1",
+                parent_hash: "0xe49352ee4984694d88eb3c1493a33d69d61c786dc5b0a32c4b3978d4fad64379",
+                nonce: "0x7622c91cd47ca43fa63f7db0ee0fd3ef",
+                timestamp: "0x16d7ad5d9de",
+                transactions_root: "0x29c04a85c4b686ec8a78615d193d64d4416dbc428f9e4631f27c62419926110f",
+                proposals_hash: "0x0000000000000000000000000000000000000000000000000000000000000000",
+                uncles_hash: "0x0000000000000000000000000000000000000000000000000000000000000000",
+                version: "0x0",
+                epoch: "0x3e80001000000",
+                dao: "0x04d0a006e1840700df5d55f5358723001206877a0b00000000e3bad4847a0100" } }
+  end
   let(:rpc) { CKB::RPC.new }
   let(:lock_hash) { "0xd0e22f863da970a3ff51a937ae78ba490bbdcede7272d658a053b9f80e30305d" }
 
@@ -79,9 +103,9 @@ RSpec.describe CKB::RPC do
       outputs: []
     }
 
-    expect {
+    expect do
       rpc.send_transaction(tx)
-    }.to raise_error(CKB::RPCError, /:code=>-3/)
+    end.to raise_error(CKB::RPCError, /:code=>-3/)
   end
 
   it "local node info" do
@@ -144,7 +168,7 @@ RSpec.describe CKB::RPC do
   end
 
   it "set ban" do
-    params = ["192.168.0.2", "insert", 1840546800000, true, "test set_ban rpc"]
+    params = ["192.168.0.2", "insert", 1_840_546_800_000, true, "test set_ban rpc"]
     result = rpc.set_ban(*params)
     expect(result).to be nil
   end
@@ -152,5 +176,39 @@ RSpec.describe CKB::RPC do
   it "get banned addresses" do
     result = rpc.get_banned_addresses
     expect(result).not_to be nil
+  end
+
+  context "miner RPCs" do
+    it "get_block_template" do
+      result = rpc.get_block_template
+      expect(result).not_to be nil
+    end
+
+    it "get_block_template with bytes_limit" do
+      result = rpc.get_block_template(1000)
+      expect(result).not_to be nil
+    end
+
+    it "get_block_template with proposals_limit" do
+      result = rpc.get_block_template(1000)
+      expect(result).not_to be nil
+    end
+
+    it "get_block_template with max_version" do
+      result = rpc.get_block_template(1000)
+      expect(result).not_to be nil
+    end
+
+    it "get_block_template with bytes_limit, proposals_limit and max_version" do
+      result = rpc.get_block_template(1000)
+      expect(result).not_to be nil
+    end
+
+    # must use real data
+    it "submit_block" do
+      raw_block_h[:header][:parent_hash] = rpc.genesis_block_hash
+      result = rpc.submit_block("test", raw_block_h)
+      expect(result).not_to be nil
+    end
   end
 end